requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]